Restructure STT engines into modular directory structure
Reorganize STT engine components from flat structure to cohesive module hierarchy. Each engine is now self-contained in its own package under eloGraf/engines/.

Changes:
- Create engines/ directory with subdirectories for each engine (nerd, whisper, google, openai, assemblyai)
- Move controller files to engines/{engine}/controller.py
- Move plugin definitions to engines/{engine}/engine.py
- Remove old plugins/ directory
- Update all imports across codebase (controllers, tests, tray_icon)
- Update engine.py files to use local imports from .controller

Benefits:
- High cohesion: All code for an engine resides in one place
- Clear structure: engines/ accurately describes functionality
- Easy navigation: Self-documenting file organization
- Scalability: Adding new engines is clean and doesn't clutter other code

All 117 tests passing.
