+++ eloGraf/engines/whisper/controller.py
